% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_raster.R
\name{as_raster}
\alias{as_raster}
\alias{as_raster,geom-method}
\title{Transform a spatial object to class \code{Raster*}}
\usage{
as_raster(x, ...)

\S4method{as_raster}{geom}(x = NULL)
}
\arguments{
\item{x}{the object to transform to class \code{Raster*}.}

\item{...}{other arguments.}
}
\value{
an object of class \code{Raster*}
}
\description{
Transform a spatial object to class \code{Raster*}
}
\seealso{
Other spatial classes: 
\code{\link{as_geometr}()},
\code{\link{as_sf}()},
\code{\link{as_sp}()},
\code{\link{as_terra}()}
}
\concept{spatial classes}
