% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{as_sf}
\alias{as_sf}
\alias{as_sf,geom-method}
\title{Transform a spatial object to class \code{sf}}
\usage{
as_sf(x, ...)

\S4method{as_sf}{geom}(x = NULL)
}
\arguments{
\item{x}{the object to transform to class \code{sf}.}

\item{...}{other arguments.}
}
\value{
If \code{x} is a \code{geom} and has attributes other than
\code{fid} and \code{gid}, a "Simple feature collection", otherwise a
"Geometry set". Several features of the \code{geom} are returned as MULTI*
feature, when they have \code{gid} and optionally other attributes in
common, otherwise they are returned as a single simple feature.
}
\description{
Transform a spatial object to class \code{sf}
}
\seealso{
Other spatial classes: 
\code{\link{as_geometr}()},
\code{\link{as_raster}()},
\code{\link{as_sp}()},
\code{\link{as_terra}()}
}
\concept{spatial classes}
