% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sp.R
\name{as_sp}
\alias{as_sp}
\alias{as_sp,geom-method}
\title{Transform a spatial object to class \code{Spatial}}
\usage{
as_sp(x, ...)

\S4method{as_sp}{geom}(x = NULL)
}
\arguments{
\item{x}{the object to transform to class \code{Spatial}.}

\item{...}{other arguments.}
}
\value{
an object of class \code{Spatial}
}
\description{
Transform a spatial object to class \code{Spatial}
}
\seealso{
Other spatial classes: 
\code{\link{as_geometr}()},
\code{\link{as_raster}()},
\code{\link{as_sf}()},
\code{\link{as_terra}()}
}
\concept{spatial classes}
