% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_terra.R
\name{as_terra}
\alias{as_terra}
\alias{as_terra,geom-method}
\title{Transform a spatial object to class \code{SpatRaster}}
\usage{
as_terra(x, ...)

\S4method{as_terra}{geom}(x = NULL)
}
\arguments{
\item{x}{the object to transform to class \code{SpatRaster}.}

\item{...}{other arguments.}
}
\value{
an object of class \code{SpatRaster}
}
\description{
Transform a spatial object to class \code{SpatRaster}
}
\seealso{
Other spatial classes: 
\code{\link{as_geometr}()},
\code{\link{as_raster}()},
\code{\link{as_sf}()},
\code{\link{as_sp}()}
}
\concept{spatial classes}
