% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cellToPointsCpp}
\alias{cellToPointsCpp}
\title{Cell coordinates to data.frame (c++)}
\usage{
cellToPointsCpp(mat)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(numeric)}\cr object of which the cell coordinates
are converted to a data.frame.}
}
\value{
a data.frame of the cell coordinates
}
\description{
C++ function that converts the cell coordinates of a matrix to a data.frame.
}
