% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countCellAdjacenciesCpp}
\alias{countCellAdjacenciesCpp}
\title{Count cell adjacencies in a matrix (c++)}
\usage{
countCellAdjacenciesCpp(mat, doublecount)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(numeric)}\cr the object in which to count
adjacencies}

\item{doublecount}{\link[=logical]{logical(1)}\cr whether or not to tally up each
adjacency for both neighbouring cells, or only once.}
}
\value{
A data.frame of values*values with their adjacencies
}
\description{
C++ function that counts the number of unique values in a matrix
}
\seealso{
Other count functions: 
\code{\link{countCellEdgesCpp}()},
\code{\link{countCellValuesCpp}()}
}
\concept{count functions}
