% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countCellEdgesCpp}
\alias{countCellEdgesCpp}
\title{Count cell edges in a matrix (c++)}
\usage{
countCellEdgesCpp(mat)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(numeric)}\cr the object in which to count edges}
}
\value{
A data.frame of the unique values and their edges
}
\description{
C++ function that counts the number of edges in a matrix
}
\seealso{
Other count functions: 
\code{\link{countCellAdjacenciesCpp}()},
\code{\link{countCellValuesCpp}()}
}
\concept{count functions}
