% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countCellValuesCpp}
\alias{countCellValuesCpp}
\title{Count unique cell values in a matrix (c++)}
\usage{
countCellValuesCpp(mat)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(numeric)}\cr the object in which to count unique
values}
}
\value{
A data.frame of the unique values and their number
}
\description{
C++ function that counts the number of unique values in a matrix
}
\seealso{
Other count functions: 
\code{\link{countCellAdjacenciesCpp}()},
\code{\link{countCellEdgesCpp}()}
}
\concept{count functions}
