% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomio-package.R
\docType{package}
\name{geomio}
\alias{geomio}
\title{geomio: geometry input/output}
\description{
The geomio package ...
geomio fills this gap by providing tools
\itemize{
\item that produce an identical output for the same metadata of different
classes (via so-called getters) and
\item that use an identical input to write to various classes that
originally require different input (via so-called setters).
}
}
\seealso{
\itemize{ \item Github project:
\href{https://github.com/EhrmannS/geomio}{https://github.com/EhrmannS/geomio}
\item Report bugs:
\href{https://github.com/EhrmannS/geomio/issues}{https://github.com/EhrmannS/geomio/issues}
}
}
\author{
\strong{Maintainer, Author}: Steffen Ehrmann
\email{steffen.ehrmann@posteo.de}
}
