% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProvenance.R
\name{getProvenance}
\alias{getProvenance}
\alias{getProvenance,geom-method}
\alias{getProvenance,Raster-method}
\title{Get the history of a spatial object.}
\usage{
getProvenance(x)

\S4method{getProvenance}{geom}(x)

\S4method{getProvenance}{Raster}(x)
}
\arguments{
\item{x}{the object from which to derive the history.}
}
\value{
A list of the events that lead to \code{x}.
}
\description{
Get the history of a spatial object.
}
\seealso{
Other getters: 
\code{\link{getCRS}()},
\code{\link{getCols}()},
\code{\link{getExtent}()},
\code{\link{getFeatures}()},
\code{\link{getGroups}()},
\code{\link{getLayers}()},
\code{\link{getNames}()},
\code{\link{getPoints}()},
\code{\link{getRes}()},
\code{\link{getRows}()},
\code{\link{getType}()},
\code{\link{getWindow}()}
}
\concept{getters}
