% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getValuesCpp}
\alias{getValuesCpp}
\title{Get Matrix cell values (c++)}
\usage{
getValuesCpp(mat)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(numeric)}\cr the matrix from which to extract the values.}
}
\value{
numeric vector of the values in \code{mat}
}
\description{
C++ function that extracts the values of a matrix
}
\seealso{
Other extractor functions: 
\code{\link{sortUniqueCpp}()}
}
\concept{extractor functions}
