% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matDistanceCpp}
\alias{matDistanceCpp}
\title{Calculate a distance map (c++)}
\usage{
matDistanceCpp(mat, method)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(numeric)}\cr the matrix with foreground patches
from which to derive the distance map.}

\item{method}{\link[=character]{character(1)}\cr which of the methods
\code{"euclidean"}, \code{"manhattan"} or \code{"chessboard"} to use.}
}
\description{
C++ function that derives a distance map of foreground patches in a matrix.
}
\seealso{
Other matrix modify functions: 
\code{\link{diamondSquareCpp}()},
\code{\link{matReduceCpp}()},
\code{\link{scaleMatrixCpp}()}
}
\concept{matrix modify functions}
