% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matInGeomCpp}
\alias{matInGeomCpp}
\title{Cell (matrix) in polygon algorithm (c++)}
\usage{
matInGeomCpp(mat, geom, negative)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(numeric)}\cr object of which the cell coordinates
are tested against \code{geom}.}

\item{geom}{\link[=matrix]{matrix(numeric)}\cr matrix of the coordinates
describing the polygon, where the first column represents x coordinate
values and the second column y coordinate values.}

\item{negative}{\link[=logical]{logical(1)}\cr ...}
}
\value{
a numeric matrix
}
\description{
C++ function that determine whether the cells in a matrix are inside or
outside of a polygon.
}
\details{
Copyright 2000 softSurfer, 2012 Dan Sunday This code may be freely
used and modified for any purpose providing that this copyright notice is
included with it. SoftSurfer makes no warranty for this code, and cannot
be held liable for any real or imagined damage resulting from its use.
Users of this code must verify correctness for their application.
}
\seealso{
Other topological relationships: 
\code{\link{pointInPolyCpp}()}
}
\concept{topological relationships}
