% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matReduceCpp}
\alias{matReduceCpp}
\title{"Reduce" a list of matrices (c++)}
\usage{
matReduceCpp(lMat, f)
}
\arguments{
\item{lMat}{\link[=list]{list(matrix)}\cr list of matrices with equal dimensions.}

\item{f}{\link[=function]{function(1)}\cr binary function to combine the matrices.}
}
\value{
a single matrix of the same dimensions as all the input matrices.
}
\description{
C++ function that uses a binary function that successively combines a list
of matrices.
}
\seealso{
Other matrix modify functions: 
\code{\link{diamondSquareCpp}()},
\code{\link{matDistanceCpp}()}
}
\concept{matrix modify functions}
