% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleMatrixCpp}
\alias{scaleMatrixCpp}
\title{... (c++)}
\usage{
scaleMatrixCpp(mat, range)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(numeric)}\cr ...}

\item{range}{\link[=numeric]{numeric(.)}\cr ...}
}
\value{
...
}
\description{
C++ function that ...
}
\seealso{
Other matrix modify functions: 
\code{\link{diamondSquareCpp}()},
\code{\link{matDistanceCpp}()},
\code{\link{matReduceCpp}()}
}
\concept{matrix modify functions}
