% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGroups.R
\name{setGroups}
\alias{setGroups}
\alias{setGroups,geom-method}
\alias{setGroups,RasterLayer-method}
\title{Set a table of group attributes}
\usage{
setGroups(x, table)

\S4method{setGroups}{geom}(x, table = NULL)

\S4method{setGroups}{RasterLayer}(x, table = NULL)
}
\arguments{
\item{x}{the object to which to assign a new attribute table.}

\item{table}{\link[=data.frame]{data.frame(1)}\cr the new attribute table.}
}
\value{
The object \code{x} with an updated group attribute table.
}
\description{
Set a table of group attributes
}
\seealso{
Other setters: 
\code{\link{setCRS}()},
\code{\link{setFeatures}()},
\code{\link{setNames}()},
\code{\link{setProvenance}()},
\code{\link{setWindow}()}
}
\concept{setters}
