% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNames.R
\name{setNames}
\alias{setNames}
\alias{setNames,geom-method}
\title{Set the names of a spatial object.}
\usage{
setNames(x, to)

\S4method{setNames}{geom}(x, to = NULL)
}
\arguments{
\item{x}{the object for which to set a new name.}

\item{to}{\link[=character]{character(1)}\cr new name.}
}
\value{
The object \code{x} with an update name.
}
\description{
Set the names of a spatial object.
}
\seealso{
Other setters: 
\code{\link{setCRS}()},
\code{\link{setFeatures}()},
\code{\link{setGroups}()},
\code{\link{setProvenance}()},
\code{\link{setWindow}()}
}
\concept{setters}
