% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setProvenance.R
\docType{methods}
\name{setProvenance}
\alias{setProvenance}
\alias{setProvenance,geom-method}
\alias{setProvenance,RasterLayer-method}
\title{Set additional entries to the history of an object}
\usage{
setProvenance(x, to)

\S4method{setProvenance}{geom}(x, to = NULL)

\S4method{setProvenance}{RasterLayer}(x, to = NULL)
}
\arguments{
\item{x}{the object for which to set the coordinate reference system.}

\item{to}{\link[=character]{character(1)}\cr a character string describing
one step in the provenance for this object; repeat to set several steps.}
}
\value{
The object \code{x} where the history slot has been updated.
}
\description{
Set additional entries to the history of an object
}
\details{
Both, objects of class \code{geom} and \code{Raster*} have the slot
\code{@history}, which contains the provenance of that object. With
\code{setProvenance}, that provenance can be updated, based on the
modification the object has been exposed to. This happens automatically for
all geometry operations that come with \code{geometr}.
}
\seealso{
Other setters: 
\code{\link{setCRS}()},
\code{\link{setFeatures}()},
\code{\link{setGroups}()},
\code{\link{setNames}()},
\code{\link{setWindow}()}
}
\concept{setters}
