% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWindow.R
\name{setWindow}
\alias{setWindow}
\alias{setWindow,geom-method}
\title{Set the reference window of a spatial object.}
\usage{
setWindow(x, to)

\S4method{setWindow}{geom}(x, to = NULL)
}
\arguments{
\item{x}{the object for which to set a new reference window.}

\item{to}{\link[=data.frame]{data.frame(1)}\cr data.frame with columns \code{x} and
\code{y} that contain the minimum and maximum values in the respective
dimension to which the reference window shall be set, see Details.}
}
\value{
The object \code{x} with an update reference window.
}
\description{
Set the reference window of a spatial object.
}
\seealso{
Other setters: 
\code{\link{setCRS}()},
\code{\link{setFeatures}()},
\code{\link{setGroups}()},
\code{\link{setNames}()},
\code{\link{setProvenance}()}
}
\concept{setters}
