% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sortUniqueCpp}
\alias{sortUniqueCpp}
\title{Extract and sort unique elements (c++)}
\usage{
sortUniqueCpp(x)
}
\arguments{
\item{x}{\link[=numeric]{numeric(.)}\cr the vector of numeric values}
}
\value{
a numeric vector of the unique and sorted values.
}
\description{
C++ function that extract and returns a sorted vector of numeric values.
}
\seealso{
Other extractor functions: 
\code{\link{getValuesCpp}()}
}
\concept{extractor functions}
