% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subChrIntCpp}
\alias{subChrIntCpp}
\title{Replace character values with integer values (c++)}
\usage{
subChrIntCpp(mat, replace, with)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(character)}\cr the object in which to replaces
character values.}

\item{replace}{\link[=character]{character(.)}\cr values to replace.}

\item{with}{\link[=integer]{integerish(.)}\cr values that replace the old values
(must have same length as \code{replace}).}
}
\value{
an integer matrix where values in \code{replace} have been replaced
with values in \code{with}.
}
\description{
C++ function that replaces character values with integer values in a matrix.
}
\seealso{
Other substitute functions: 
\code{\link{subNumChrCpp}()},
\code{\link{subNumNumCpp}()}
}
\concept{substitute functions}
