% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomio-package.R
\docType{package}
\name{geomio}
\alias{geomio}
\title{geomio: geometry input/output}
\description{
The geomio package ...
}
\seealso{
\itemize{ \item Github project:
\href{https://github.com/EhrmannS/geomio}{https://github.com/EhrmannS/geomio}
\item Report bugs:
\href{https://github.com/EhrmannS/geomio/issues}{https://github.com/EhrmannS/geomio/issues}
}
}
\author{
\strong{Maintainer, Author}: Steffen Ehrmann
\email{steffen.ehrmann@posteo.de}
}
